consolidate project query/fetcher
